requirements.txt
absl-py==0.1.13
alabaster==0.7.10
astor==0.6.2
Babel==2.5.3
bcolz==1.1.2
beautifulsoup4==4.6.0
bleach==1.5.0
boto==2.48.0
boto3==1.5.22
botocore==1.8.36
bz2file==0.98
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
cymem==1.31.2
cytoolz==0.8.2
decorator==4.2.1
dill==0.2.7.1
docutils==0.14
elasticsearch==6.1.1
entrypoints==0.2.3
enum34==1.1.6
ftfy==4.4.3
gast==0.2.0
gensim==3.2.0
graphviz==0.8.2
grpcio==1.10.1
h5py==2.7.1
html5lib==0.9999999
idna==2.6
imagesize==0.7.1
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
kaggle==1.1.0
Keras==2.1.5
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.1
mistune==0.8.3
msgpack-numpy==0.4.1
msgpack-python==0.4.8
murmurhash==0.28.0
mxnet==1.0.0.post4
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.14.2
packaging==17.1
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pathlib==1.0.1
pickleshare==0.7.4
Pillow==5.0.0
plac==0.9.6
preshed==1.0.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
py4j==0.10.4
Pygments==2.2.0
pymongo==3.6.0
pyparsing==2.2.0
pyreadline==2.1
pyspark==2.2.1
python-dateutil==2.6.1
pytz==2017.3
pywinpty==0.5.1
PyYAML==3.12
pyzmq==16.0.4
qtconsole==4.3.1
regex==2017.4.5
requests==2.18.4
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==1.0.1
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
smart-open==1.5.6
snowballstemmer==1.2.1
spacy==2.0.5
Sphinx==1.7.2
sphinxcontrib-websupport==1.0.1
tensorboard==1.7.0
tensorflow==1.7.0
tensorflow-tensorboard==0.4.0rc3
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
Theano==1.0.1
thinc==6.10.2
toolz==0.8.2
tornado==4.5.3
tqdm==4.19.5
traitlets==4.3.2
ujson==1.35
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.3
wrapt==1.10.11
